debugged the platforms (mostly)
hi guys,

I made some progress here. I got the platforms to work individually in a string of if, else if, else statements. And I removed jumpable. It was redundant to have grounded and jumpable (I think), because you should only be able to jump if you're grounded, right? Anyway, I couldn't get it to work on the dynamic platforms, but you know your code better, so maybe you can. Good luck!!!